+++ app/components/hero/hero.tsx
@@ -4,7 +4,7 @@ import { motion } from 'framer-motion';
 import Link from 'next/link';
 import { useContext, useEffect, useRef } from 'react';
 import { HiOutlineArrowNarrowDown } from 'react-icons/hi';
-import { ScrollContext } from '../../../components/Providers/ScrollProvider';
+import { ScrollContext } from '../providers/ScrollProvider';
 import { renderCanvas } from './renderCanvas';
 
 export default function Hero() {
