+++ app/projects/project-item.tsx
