+++ app/projects/project-preview.tsx
